backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6

# Database and ORM
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis for caching
redis==5.0.1
aioredis==2.0.1

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8

# Audio Processing
librosa==0.10.1
soundfile==1.0.0
webrtcvad==2.0.10
pydub==0.25.1

# AI/ML Libraries
openai==1.3.5
whisper==1.1.10
torch==2.1.1
torchaudio==2.1.1
transformers==4.35.2
numpy==1.24.3
scipy==1.11.4

# Anti-Spoofing Models
speechbrain==0.5.16

# HTTP Client
httpx==0.25.2
aiofiles==23.2.1

# Validation and Parsing
pydantic==2.5.0
email-validator==2.1.0

# Environment and Configuration
python-dotenv==1.0.0

# Monitoring and Logging
prometheus-client==0.19.0
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Data Serialization
orjson==3.9.10

# Rate Limiting
slowapi==0.1.9

# Background Tasks
celery==5.3.4
